Integration with ns-3.28 (not completed yet)

+++ src/uan/model/uan-channel.cc
@@ -27,8 +27,6 @@
 #include "ns3/log.h"
 #include "ns3/pointer.h"
 #include "ns3/string.h"
-#include "ns3/log.h"
-
 #include "uan-channel.h"
 #include "uan-phy.h"
 #include "uan-prop-model.h"
@@ -129,7 +127,7 @@ UanChannel::SetPropagationModel (Ptr<UanPropModel> prop)
 uint32_t
 UanChannel::GetNDevices () const
 {
-  return m_devList.size ();
+  return static_cast<uint32_t> (m_devList.size ());
 }
 
 Ptr<NetDevice>
